requirements.txt
streamlit==1.40.2
pandas==2.2.3
gspread==6.1.4
google-auth==2.35.0
